pyproject.toml
[tool.poetry]
name = "brainseg"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
svgutils = "^0.3.4"
ipykernel = "^6.13.0"
Pillow = "^9.1.0"
numpy = "^1.22.3"
tqdm = "^4.64.0"
openslide-python = "^1.1.2"
streamlit = "^1.9.0"
matplotlib = "^3.5.2"
czifile = "^2019.7.2"
scikit-learn = "^1.0.2"
scikit-image = "^0.19.2"
aicspylibczi = "^3.0.5"
tensorflow = "2.8.0"
segmentation-models = "^1.0.1"
pypotrace = "^0.3"
imgaug = "^0.4.0"
pydot = "^1.4.2"
traitlets = "5.1.1"
notebook = "^6.4.12"
nibabel = "^4.0.1"
nilearn = "^0.9.2"
silence-tensorflow = "^1.2.1"
Shapely = "^1.8.4"
SimpleITK = "^2.2.0"
statsmodels = "^0.13.2"
numba = "^0.56.3"
xarray = "^2022.10.0"
antspyx = "^0.3.5"
geojson = "^2.5.0"
imagecodecs = "^2023.3.16"
geopandas = "^0.12.2"
itk-elastix = "^0.16.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
process_svg_for_registration = "scripts.process_svg_for_registration:main"
